+++ apps/Changelog.md
@@ -3,10 +3,11 @@
 Feature:
 - [ ] set redis server store mode.
 
+- [x] support hd video address
 - [x] support the user system
   - [x] login/logout
   - [x] save to the redis, use the hash `userinfo` include the fields: `username, nickname, email, password, salt, createAt, loginAt, loginIp`
   - [x] register
 
-- [ ] favorite list
-  - [ ] save the users' favorite list, use the hash `favorite` include the fields: `username, createAt, vid, title, cover`
+- [x] favorite list
+  - [x] save the users' favorite list, use the hash `favorite:$(username)` include the fields: `username, createAt, vid, title, cover, href`